Correction des répertoires des backups.

+++ core/modules/BackupManager/Installer.php
@@ -14,11 +14,30 @@ class Installer implements \SoosyzeCore\System\Migration
     public function install(ContainerInterface $ci)
     {
         /* Création du dossier de backup. */
-        $dir = $ci->core()->getDir('backup_dir', '../soosyze_backups');
-        if (!file_exists($dir)) {
-            \mkdir($dir, 644, true);
+        $dir = $ci->core()->getDir('backup_dir', '../soosyze_backups/default');
+        if (!is_dir($dir)) {
+            \mkdir($dir, 0755, true);
+            $f = fopen($dir . '/.htaccess', 'w');
+            fwrite($f, '
+# Apache 2.4+.
+<IfModule mod_authz_core.c>
+  Require all denied
+</IfModule>
+
+# Apache 2.0 & 2.2.*
+<IfModule !mod_authz_core.c>
+  Deny from all
+</IfModule>
+
+Options None
+Options +FollowSymLinks
+
+<IfModule mod_php5.c>
+  php_flag engine off
+</IfModule>');
+            fclose($f);
         }
-        
+
         $ci->config()
             ->set('settings.max_backups', 30)
             ->set('settings.backup_cron', false);
@@ -27,13 +46,13 @@ class Installer implements \SoosyzeCore\System\Migration
     public function seeders(ContainerInterface $ci)
     {
     }
-    
+
     public function hookInstall(ContainerInterface $ci)
     {
         $this->hookInstallMenu($ci);
         $this->hookInstallUser($ci);
     }
-    
+
     public function hookInstallUser(ContainerInterface $ci)
     {
         if ($ci->module()->has('User')) {
@@ -49,9 +68,11 @@ class Installer implements \SoosyzeCore\System\Migration
         if ($ci->module()->has('Menu')) {
             $ci->query()
                 ->insertInto('menu_link', [
-                    'key', 'icon', 'title_link', 'link', 'menu', 'weight', 'parent', 'active'
+                    'key', 'icon', 'title_link', 'link', 'menu', 'weight', 'parent',
+                    'active'
                 ])
-                ->values([ 'backupmanager.index', 'fas fa-file-archive', 'Backups', 'admin/backupmanager/backups', 'menu-admin', 50, -1, true ])
+                ->values([ 'backupmanager.index', 'fas fa-file-archive', 'Backups',
+                    'admin/backupmanager/backups', 'menu-admin', 50, -1, true ])
                 ->execute();
         }
     }
@@ -61,7 +82,7 @@ class Installer implements \SoosyzeCore\System\Migration
         $this->hookUninstallMenu($ci);
         $this->hookUninstallUser($ci);
     }
-    
+
     public function hookUninstallMenu(ContainerInterface $ci)
     {
         if ($ci->module()->has('Menu')) {
@@ -71,7 +92,7 @@ class Installer implements \SoosyzeCore\System\Migration
                 ->execute();
         }
     }
-    
+
     public function hookUninstallUser(ContainerInterface $ci)
     {
         if ($ci->module()->has('User')) {
@@ -85,13 +106,17 @@ class Installer implements \SoosyzeCore\System\Migration
 
     public function uninstall(ContainerInterface $ci)
     {
+        $dir = $ci->core()->getDir('backup_dir', '../soosyze_backups');
+        if (!is_dir($dir)) {
+            return;
+        }
         /* Suppression du dossier de sauvegarde. */
-        foreach (new \DirectoryIterator($ci->core()->getDir('backup_dir')) as $file) {
+        foreach (new \DirectoryIterator($dir) as $file) {
             if ($file->isDot()) {
                 continue;
             }
             \unlink($file->getPathname());
         }
-        rmdir($ci->core()->getDir('backup_dir', '../soosyze_backups'));
+        rmdir($dir);
     }
 }
